requirements.txt
Pillow
Open-Tamil
